[Feat] 더미데이터 적용
